requirements.txt
requests
beautifulsoup4
apache-airflow-providers-http
apache-airflow-providers-postgres